+++ src/main/java/com/telepathicgrunt/repurposedstructures/configs/RSShipwrecksConfig.java
@@ -57,7 +57,7 @@ public class RSShipwrecksConfig implements ConfigData {
         @Comment("How rare are Nether Bricks Shipwreck in any non-warped or non-crimson Nether biome."
                 + "\n1 for spawning in most chunks and 1001 for none.")
         @ConfigEntry.BoundedDiscrete(min = 1, max = 1001)
-        public int netherBricksShipwreckSpawnrate = 55;
+        public int netherBricksShipwreckSpawnrate = 51;
 
 
         @ConfigEntry.Gui.Tooltip(count = 0)
@@ -65,7 +65,7 @@ public class RSShipwrecksConfig implements ConfigData {
         @Comment("How rare are Crimson Shipwreck in Crimson Nether biome."
                 + "\n1 for spawning in most chunks and 1001 for none.")
         @ConfigEntry.BoundedDiscrete(min = 1, max = 1001)
-        public int crimsonShipwreckSpawnrate = 43;
+        public int crimsonShipwreckSpawnrate = 41;
 
 
         @ConfigEntry.Gui.Tooltip(count = 0)
@@ -73,6 +73,6 @@ public class RSShipwrecksConfig implements ConfigData {
         @Comment("How rare are Warped Shipwreck in Warped Nether biome."
                 + "\n1 for spawning in most chunks and 1001 for none.")
         @ConfigEntry.BoundedDiscrete(min = 1, max = 1001)
-        public int warpedShipwreckSpawnrate = 43;
+        public int warpedShipwreckSpawnrate = 41;
     }
 }