+++ src/main/java/com/telepathicgrunt/repurposedstructures/world/structures/VillageNetherStructure.java
@@ -34,7 +34,7 @@ public class VillageNetherStructure extends VillageBaseStructure {
 
         for (int curChunkX = chunkX - 10; curChunkX <= chunkX + 10; curChunkX++) {
             for (int curChunkZ = chunkZ - 10; curChunkZ <= chunkZ + 10; curChunkZ++) {
-                for(StructureFeature<DefaultFeatureConfig> outpost : RSStructures.NETHER_OUTPOSTS_LIST){
+                for(StructureFeature<?> outpost : RSStructures.NETHER_OUTPOSTS_LIST){
                     ChunkPos chunkPos2 = outpost.getStartChunk(Objects.requireNonNull(chunkGenerator.getStructuresConfig().getForType(outpost)), seed, chunkRandom, curChunkX, curChunkZ);
                     if (curChunkX == chunkPos2.x && curChunkZ == chunkPos2.z) {
                         return false;